Fixed auto detection of unpaired library

+++ libexec/getreadsSRAonBW.sh
@@ -3,8 +3,13 @@ accession=$1
 module load perl
 module load sratoolkit
 fastq-dump --split-e "$accession"
-mv "$accession"_1.fastq "$accession"_R1.fastq
-mv "$accession"_2.fastq "$accession"_R2.fastq
-rm "$accession".fastq
+splitfile="$accession"_1.fastq
+if [ -f "$splitfile" ]
+then
+    mv "$accession"_1.fastq "$accession"_R1.fastq
+    mv "$accession"_2.fastq "$accession"_R2.fastq
+    rm "$accession".fastq
+fi
+
 module unload sratoolkit
 module unload perl
